+++ source/common/common/cleanup.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <functional>
+#include <memory>
 
 namespace Envoy {
 
@@ -13,5 +14,7 @@ public:
 private:
   std::function<void()> f_;
 };
+
+typedef std::unique_ptr<Cleanup> CleanupPtr;
 
 } // namespace Envoy
