+++ source/common/http/http2/nghttp2.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace Envoy {
+namespace Http {
+namespace Http2 {
+
+/**
+ * Setup nghttp2 trace-level logging for when debugging.
+ */
+void initializeNghttp2Logging();
+
+} // namespace Http2
+} // namespace Http
+} // namespace Envoy
