+++ source/common/network/dns_impl.cc
@@ -25,11 +25,6 @@ DnsResolverImpl::DnsResolverImpl(
     const std::vector<Network::Address::InstanceConstSharedPtr>& resolvers)
     : dispatcher_(dispatcher),
       timer_(dispatcher.createTimer([this] { onEventCallback(ARES_SOCKET_BAD, 0); })) {
-  // This is also done in main(), to satisfy the requirement that c-ares is
-  // initialized prior to threading. The additional call to ares_library_init()
-  // here is a nop in normal execution, but exists for testing where we don't
-  // launch via main().
-  ares_library_init(ARES_LIB_INIT_ALL);
   ares_options options;
 
   initializeChannel(&options, 0);
@@ -63,7 +58,6 @@ DnsResolverImpl::DnsResolverImpl(
 DnsResolverImpl::~DnsResolverImpl() {
   timer_->disableTimer();
   ares_destroy(channel_);
-  ares_library_cleanup();
 }
 
 void DnsResolverImpl::initializeChannel(ares_options* options, int optmask) {
