+++ test/BUILD
@@ -3,7 +3,6 @@ licenses(["notice"])  # Apache 2
 load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_test_library",
-    "envoy_google_grpc_external_deps",
     "envoy_package",
 )
 
@@ -31,6 +30,7 @@ envoy_cc_test_library(
         "//source/common/common:thread_lib",
         "//source/common/event:libevent_lib",
         "//source/common/http/http2:codec_lib",
+        "//source/exe:process_wide_lib",
         "//test/common/runtime:utility_lib",
         "//test/mocks/access_log:access_log_mocks",
         "//test/test_common:environment_lib",
@@ -39,5 +39,5 @@ envoy_cc_test_library(
     ] + select({
         "//bazel:disable_signal_trace": [],
         "//conditions:default": ["//source/exe:sigaction_lib"],
-    }) + envoy_google_grpc_external_deps(),
+    }),
 )
