+++ test/fuzz/fuzz_runner.cc
@@ -5,6 +5,8 @@
 #include "common/event/libevent.h"
 #include "common/http/http2/codec_impl.h"
 
+#include "exe/process_wide.h"
+
 #include "test/test_common/environment.h"
 
 namespace Envoy {
@@ -24,9 +26,9 @@ PerTestEnvironment::PerTestEnvironment()
 PerTestEnvironment::~PerTestEnvironment() { TestEnvironment::removePath(test_tmpdir_); }
 
 void Runner::setupEnvironment(int argc, char** argv, spdlog::level::level_enum default_log_level) {
-  Event::Libevent::Global::initialize();
-  Http::Http2::initializeNghttp2Logging();
-
+  // We hold on to process_cleanup to provide RAII cleanup of process-wide
+  // state.
+  auto process_cleanup = Envoy::ProcessWide::setup();
   TestEnvironment::initializeOptions(argc, argv);
 
   const auto environment_log_level = TestEnvironment::getOptions().logLevel();
