+++ test/test_runner.h
@@ -7,16 +7,14 @@
 #include "common/http/http2/codec_impl.h"
 #include "common/runtime/runtime_features.h"
 
+#include "exe/process_wide.h"
+
 #include "test/common/runtime/utility.h"
 #include "test/mocks/access_log/mocks.h"
 #include "test/test_common/environment.h"
 #include "test/test_listener.h"
 
 #include "gmock/gmock.h"
-
-#ifdef ENVOY_GOOGLE_GRPC
-#include "grpc/grpc.h"
-#endif
 
 namespace Envoy {
 namespace {
@@ -69,12 +67,9 @@ class TestRunner {
 public:
   static int RunTests(int argc, char** argv) {
     ::testing::InitGoogleMock(&argc, argv);
-#ifdef ENVOY_GOOGLE_GRPC
-    grpc_init();
-#endif
-    Event::Libevent::Global::initialize();
-    Http::Http2::initializeNghttp2Logging();
-
+    // We hold on to process_cleanup to provide RAII cleanup of process-wide
+    // state.
+    auto process_cleanup = ProcessWide::setup();
     // Add a test-listener so we can call a hook where we can do a quiescence
     // check after each method. See
     // https://github.com/google/googletest/blob/master/googletest/docs/advanced.md
